Migrate regressions from 5001-5501
