+++ spacy/tests/serialize/test_serialize_doc.py
@@ -189,6 +189,16 @@ def test_issue4528(en_vocab):
     assert new_doc.user_data[("._.", "foo", None, None)] == "bar"
 
 
+@pytest.mark.issue(5141)
+def test_issue5141(en_vocab):
+    """Ensure an empty DocBin does not crash on serialization"""
+    doc_bin = DocBin(attrs=["DEP", "HEAD"])
+    assert list(doc_bin.get_docs(en_vocab)) == []
+    doc_bin_bytes = doc_bin.to_bytes()
+    doc_bin_2 = DocBin().from_bytes(doc_bin_bytes)
+    assert list(doc_bin_2.get_docs(en_vocab)) == []
+
+
 def test_serialize_empty_doc(en_vocab):
     doc = Doc(en_vocab)
     data = doc.to_bytes()
